TODO API created on Flask.
Based on this document
https://gist.github.com/prograMind/e403103c81221a30a310f5717904def9

First version implemented everything from the doc
Added basic Functional tests.

Tasks are stored in session. But can easily be made to work with something else.

TODO Design an API change to make it possible to create a nested list of subtasks for every task
Refactor code to make the data layer (session now) more easy to change.